pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workspace-secretary"
version = "0.1.0"
description = "Google Workspace Secretary MCP server for interactive email and calendar processing"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "John Neerdael", email = "example@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp>=0.1.0",
    "imapclient>=2.2.0",
    "email-validator>=1.1.3",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.3",
    "requests>=2.32.3",
    "google-auth>=2.38.0",
    "ruff>=0.11.2",
    "google-auth-oauthlib>=1.2.1",
    "flask>=3.1.0",
    "uvicorn>=0.20.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.19.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "mypy>=0.982",
    "flake8>=5.0.0",
    "pre-commit>=2.19.0",
]

[project.urls]
Homepage = "https://github.com/johnneerdael/Google-Workspace-Secretary-MCP"
Issues = "https://github.com/johnneerdael/Google-Workspace-Secretary-MCP/issues"

[tool.setuptools]
packages = ["workspace_secretary"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
